backend/requirements.txt
fastapi
uvicorn
scikit-learn
joblib
numpy
pandas
pydantic
python-multipart
httpx
firebase-admin
python-dateutil
razorpay
resend
python-dotenv
pillow
pdf2image
pytesseract
